calendar: Copy Source from MessageTray
The new notification system will still use the concept of notification
sources, but heavily deemphasized and not as the primary entry point
of the notification system. Therefore it is much easier to put the new
system in place and move over once it's done than to modify the existing
code gradually without breaking consumers.