messageTray: Make Notification UI-less
We no longer have a single notification actor that is either displayed
as banner or moved to the summary, depending on state - both the lock
screen and the message list simply use the information from the object
to create their own UI representation. It doesn't really make sense to
make banners special there, in particular as its usually not the
longest-lived representation. So remove any actual UI from the
Notification class itself, and use a separate NotificationBanner
which builds on the existing MessageList actors.